RENAME [LONG,DOUBLE,STRING] -> [INTEGER,FLOAT,STRING]

+++ src/main/java/ru/cft/drozdetskiy/ContentType.java
@@ -2,17 +2,17 @@ package ru.cft.drozdetskiy;
 
 /**
  * Можно распарсить строку в целое число, либо в вещественное число, либо оставить строкой.
- * Типы содержимого строки: {@linkplain #LONG}, {@linkplain #DOUBLE}, {@linkplain #STRING}
+ * Типы содержимого строки: {@linkplain #INTEGER}, {@linkplain #FLOAT}, {@linkplain #STRING}
  */
 enum ContentType {
     /**
      * Целое число.
      */
-    LONG,
+    INTEGER,
     /**
      * Вещественное число.
      */
-    DOUBLE,
+    FLOAT,
     /**
      * Строка.
      */
